substituted VP with submodule
