+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "v2/Volume-Profiles-Spicetify-Extension-v2"]
+	path = v2/Volume-Profiles-Spicetify-Extension-v2
+	url = https://github.com/notPlancha/Volume-Profiles-Spicetify-Extension-v2.git
