+++ v2/volume-profiles/.gitignore
@@ -142,3 +142,5 @@ out
 .svelte-kit
 
 # End of https://www.toptal.com/developers/gitignore/api/node
+
+.idea/
