+++ v2/volume-profiles/README.md
@@ -1,5 +1,9 @@
 # Volume Profiles v2
 
+###### Built with [Spicetify Creator](https://github.com/FlafyDev/spicetify-creator). Download from [Spicetify Marketplace](https://github.com/CharlieS1103/spicetify-marketplace).
+
+[![repo](https://img.shields.io/github/last-commit/notPlancha/Volume-Profiles-Spicetify-Extension-v2)](https://github.com/notPlancha/Volume-Profiles-Spicetify-Extension-v2)
+
 This is the upgraded version of [Volume-Profiles-Spicetify-Extension-v1](https://github.com/notPlancha/Volume-Profiles-Spicetify-Extension-v1), where it's more responsive, more extended, and more Spicy.
 
 This exension is an extension for Spicetify-cli to save a volume to a profile, to load later when needed.
@@ -12,22 +16,28 @@ npm i
 
 ## Preview
 
-![buttons](https://github.com/notPlancha/Volume-Profiles-Spicetify-Extension-v2/blob/main/assets/buttons.png)
-![settings](https://github.com/notPlancha/Volume-Profiles-Spicetify-Extension-v2/blob/main/assets/settings.png)
+![buttons](assets/buttons.png)
+![settings](assets/settings.png)
 
 ## Notes
 
 * This package uses uzairfarooq's [arrive.js](https://github.com/uzairfarooq/arrive) and franciscop's [umbrella js](https://github.com/franciscop/umbrella)
 
-* It's possible that an update to spotify or spicetify will break your saved volumes, but I've yet to experience this. If this happens, try raising an issue on the [Volume-Profiles-Spicetify-Extension-v2](https://github.com/notPlancha/Volume-Profiles-Spicetify-Extension-v2)'s issues page
+* It's possible that an update to spotify or spicetify will break your saved volumes, but I've yet to experience this. If this happens, try raising an issue on the github page
+[create issue](https://github.com/user/repository/issues/new)
 - Same thing for any other issue/bug
 
 ## More
 
-- Check out me ([@notPlancha](https://github.com/notPlancha)) for possibly future extensions, and check the where the repository is updated and stuff: [Volume-Profiles-Spicetify-Extension-v2](https://github.com/notPlancha/Volume-Profiles-Spicetify-Extension-v2)
+- Check me out for possibly future extensions, star the repo if you liked it and watch the repo for updates
+
+[![follow me](https://img.shields.io/github/followers/notPlancha?style=social)](https://github.com/notPlancha)
+
+[![GitHub Repo stars](https://img.shields.io/github/stars/notPlancha/Volume-Profiles-Spicetify-Extension-v2?style=social)](https://github.com/notPlancha/Volume-Profiles-Spicetify-Extension-v2)
+
+[![GitHub watchers](https://img.shields.io/github/watchers/notPlancha/Volume-Profiles-Spicetify-Extension-v2?style=social)](https://github.com/user/repository/subscription)
+
 
 - If you have any suggestions, drop it on the [discussions page](https://github.com/notPlancha/Volume-Profiles-Spicetify-Extension-v2/discussions) or make a pull request
 
-## Made with Spicetify Creator
-
-- https://github.com/FlafyDev/spotify-css-editor
+- version 2.1.2
