+++ v2/volume-profiles/package.json
@@ -16,7 +16,7 @@
   "dependencies": {
     "arrive": "^2.4.1",
     "lodash": "^4.17.21",
-    "spcr-settings": "^1.0.2",
+    "spcr-settings": "^1.1.0",
     "umbrellajs": "^3.3.1"
   }
 }
