+++ v2/volume-profiles/run.bat
@@ -1,0 +1,1 @@
+call npm run build && call spicetify apply && call spicetify enable-devtools && call npm run build-local
